feat(all): readme优化 修复bug 统一身份管理系统

+++ machine-services/machine-data-service/src/main/java/com/machine/service/data/tag/service/impl/DataTagServiceImpl.java
@@ -3,16 +3,14 @@ package com.machine.service.data.tag.service.impl;
 import cn.hutool.core.collection.CollectionUtil;
 import cn.hutool.json.JSONUtil;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
-import com.machine.client.data.tag.dto.input.DataTagCreateInputDto;
-import com.machine.client.data.tag.dto.input.DataTagQueryPageInputDto;
-import com.machine.client.data.tag.dto.input.DataTagUpdateCategoryDto;
-import com.machine.client.data.tag.dto.input.DataTagUpdateInputDto;
+import com.machine.client.data.tag.dto.input.*;
 import com.machine.client.data.tag.dto.output.DataTagDetailOutputDto;
 import com.machine.client.data.tag.dto.output.DataTagListOutputDto;
 import com.machine.sdk.common.exception.data.DataBusinessException;
 import com.machine.sdk.common.model.request.IdRequest;
 import com.machine.service.data.tag.dao.IDataTagDao;
 import com.machine.service.data.tag.dao.IDataTagCategoryDao;
+import com.machine.service.data.tag.dao.IDataTagOptionDao;
 import com.machine.service.data.tag.dao.mapper.entity.DataTagEntity;
 import com.machine.service.data.tag.dao.mapper.entity.DataTagCategoryEntity;
 import com.machine.service.data.tag.service.IDataTagService;
@@ -37,6 +35,9 @@ public class DataTagServiceImpl implements IDataTagService {
     @Autowired
     private IDataTagCategoryDao tagCategoryDao;
 
+    @Autowired
+    private IDataTagOptionDao dataTagOptionDao;
+
     @Override
     @Transactional(rollbackFor = Exception.class)
     public String create(DataTagCreateInputDto inputDto) {
@@ -44,6 +45,12 @@ public class DataTagServiceImpl implements IDataTagService {
         DataTagCategoryEntity categoryEntity = tagCategoryDao.getById(inputDto.getCategoryId());
         if (categoryEntity == null) {
             throw new DataBusinessException("data.tag.service.create.categoryIdNotExists", "分类ID不存在");
+        }
+
+        // 验证编码是否已存在
+        DataTagEntity entityByCode = tagDao.getByCode(inputDto.getCode());
+        if (entityByCode != null) {
+            throw new DataBusinessException("data.tag.service.create.codeAlreadyExists", "标签编码已存在");
         }
 
         // 验证名称是否已存在
@@ -54,15 +61,26 @@ public class DataTagServiceImpl implements IDataTagService {
 
         DataTagEntity entity = new DataTagEntity();
         entity.setCategoryId(inputDto.getCategoryId());
-        entity.setSort(inputDto.getSort());
+        entity.setCode(inputDto.getCode());
         entity.setName(inputDto.getName());
-
+        entity.setSort(inputDto.getSort());
+        entity.setDescription(inputDto.getDescription());
         return tagDao.insert(entity);
     }
 
     @Override
     @Transactional(rollbackFor = Exception.class)
     public int delete(IdRequest request) {
+        DataTagCategoryEntity entity = tagCategoryDao.getById(request.getId());
+        if (null == entity) {
+            return 0;
+        }
+
+        //判断智能标签分类是否关联标签选项
+        if (CollectionUtil.isNotEmpty(dataTagOptionDao.selectByTagId(entity.getId()))) {
+            throw new DataBusinessException("data.tagCategory.service.delete.associationTagOption", "关联标签选项不能删除");
+        }
+
         return tagDao.delete(request.getId());
     }
 
@@ -82,8 +100,71 @@ public class DataTagServiceImpl implements IDataTagService {
 
         DataTagEntity entity = new DataTagEntity();
         entity.setId(inputDto.getId());
-        entity.setSort(inputDto.getSort());
         entity.setName(inputDto.getName());
+        entity.setDescription(inputDto.getDescription());
+        return tagDao.update(entity);
+    }
+
+    @Override
+    @Transactional(rollbackFor = Exception.class)
+    public int updateCode(DataTagUpdateCodeInputDto inputDto) {
+        DataTagEntity dbEntity = tagDao.getById(inputDto.getId());
+        if (dbEntity == null) {
+            return 0;
+        }
+
+        if (dbEntity.getCode().equals(inputDto.getCode())) {
+            //相同直接返回
+            return 0;
+        }
+
+        // 验证编码是否已存在
+        DataTagEntity entityByCode = tagDao.getByCode(inputDto.getCode());
+        if (entityByCode != null && !entityByCode.getId().equals(inputDto.getId())) {
+            throw new DataBusinessException("data.tag.service.updateCode.codeAlreadyExists", "标签编码已存在");
+        }
+
+        DataTagEntity entity = new DataTagEntity();
+        entity.setId(inputDto.getId());
+        entity.setCode(inputDto.getCode());
+        return tagDao.update(entity);
+    }
+
+    @Override
+    @Transactional(rollbackFor = Exception.class)
+    public int updateStatus(DataTagUpdateStatusInputDto inputDto) {
+        DataTagEntity dbEntity = tagDao.getById(inputDto.getId());
+        if (dbEntity == null) {
+            return 0;
+        }
+
+        if (dbEntity.getStatus().equals(inputDto.getStatus())) {
+            //相同直接返回
+            return 0;
+        }
+
+        DataTagEntity entity = new DataTagEntity();
+        entity.setId(inputDto.getId());
+        entity.setStatus(inputDto.getStatus());
+        return tagDao.update(entity);
+    }
+
+    @Override
+    @Transactional(rollbackFor = Exception.class)
+    public int updateSort(DataTagUpdateSortInputDto inputDto) {
+        DataTagEntity dbEntity = tagDao.getById(inputDto.getId());
+        if (dbEntity == null) {
+            return 0;
+        }
+
+        if (dbEntity.getSort().equals(inputDto.getSort())) {
+            //相同直接返回
+            return 0;
+        }
+
+        DataTagEntity entity = new DataTagEntity();
+        entity.setId(inputDto.getId());
+        entity.setSort(inputDto.getSort());
         return tagDao.update(entity);
     }
